% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as.data.frame.retraction_result}
\alias{as.data.frame.retraction_result}
\title{Coerce a retraction result to a data frame.}
\usage{
\method{as.data.frame}{retraction_result}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{retraction_result}.}

\item{row.names, optional}{Passed to \code{\link[base:as.data.frame]{base::as.data.frame()}}.}

\item{...}{Passed to \code{\link[base:as.data.frame]{base::as.data.frame()}}.}
}
\description{
Coerce a retraction result to a data frame.
}
