% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as_tibble.retraction_result}
\alias{as_tibble.retraction_result}
\title{Coerce a retraction result to a plain tibble.}
\usage{
\method{as_tibble}{retraction_result}(x, ...)
}
\arguments{
\item{x}{A \code{retraction_result}.}

\item{...}{Unused.}
}
\description{
Coerce a retraction result to a plain tibble.
}
