% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{list_backends}
\alias{list_backends}
\title{Names of all registered backends.}
\usage{
list_backends()
}
\value{
A character vector of source names.
}
\description{
Names of all registered backends.
}
\examples{
list_backends()
}
