% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.retraction_result}
\alias{print.retraction_result}
\title{Print a retraction result.}
\usage{
\method{print}{retraction_result}(x, ...)
}
\arguments{
\item{x}{A \code{retraction_result}.}

\item{...}{Unused.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a retraction result.
}
