% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{retraction_cache_dir}
\alias{retraction_cache_dir}
\title{Location of the retraction cache directory}
\usage{
retraction_cache_dir(create = FALSE)
}
\arguments{
\item{create}{Create the directory if it does not exist.}
}
\value{
The cache directory path (via \code{\link[tools:R_user_dir]{tools::R_user_dir()}}).
}
\description{
Location of the retraction cache directory
}
\examples{
retraction_cache_dir()
}
