% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{retraction_clear_cache}
\alias{retraction_clear_cache}
\title{Delete the local snapshot and clear the in-memory cache}
\usage{
retraction_clear_cache()
}
\value{
Invisibly \code{TRUE}.
}
\description{
Delete the local snapshot and clear the in-memory cache
}
\examples{
\dontrun{
retraction_clear_cache()
}
}
